Make max and min search periods configurable.
The default max search period for the channel access client in EPICS base
is 300 seconds, and it honours the environment variable
EPICS_CA_MAX_SEARCH_PERIOD. The default for JCA is 30 seconds. This commit
makes that value (and the equivalent min value) configurable either using
the environment variable or the jca properties file.